+++ R/chrome.R
@@ -43,18 +43,23 @@
 #' @param async Execute \code{chrome_print()} asynchronously? If \code{TRUE},
 #'   \code{chrome_print()} returns a \code{\link[promises]{promise}} value (the
 #'   \pkg{promises} package has to be installed in this case).
+#' @param encoding Not used. This argument is required by RStudio IDE.
 #' @references
 #' \url{https://developers.google.com/web/updates/2017/04/headless-chrome}
 #' @return Path of the output file (invisibly). If \code{async} is \code{TRUE}, this
 #'   is a \code{\link[promises]{promise}} value.
-#' @seealso \link{knit_pdf_chrome}()
 #' @export
 chrome_print = function(
   input, output = xfun::with_ext(input, format), wait = 2, browser = 'google-chrome',
   format = c('pdf', 'png', 'jpeg'), options = list(), selector = 'body',
   box_model = c('border', 'content', 'margin', 'padding'), scale = 1, work_dir = tempfile(),
-  timeout = 30, extra_args = c('--disable-gpu'), verbose = 0, async = FALSE
+  timeout = 30, extra_args = c('--disable-gpu'), verbose = 0, async = FALSE, encoding
 ) {
+  is_rstudio_knit =
+    !interactive() && !is.na(Sys.getenv('RSTUDIO', NA)) &&
+    !missing(encoding) && length(match.call()) == 3
+  if (is_rstudio_knit) verbose = 1
+
   if (missing(browser) && is.na(browser <- Sys.getenv('PAGEDOWN_CHROME', NA))) {
     browser = find_chrome()
   } else {
@@ -118,9 +123,10 @@ chrome_print = function(
     svr = NULL # init svr variable
     if (file.exists(input)) {
       is_html = function(x) grepl('[.]html?$', x)
-      url = if (is_html(input)) input else render_opt_msg(
+      with_msg_maybe = if (is_rstudio_knit) suppressMessages else identity
+      url = if (is_html(input)) input else with_msg_maybe(rmarkdown::render(
         input, envir = parent.frame(), encoding = 'UTF-8'
-      )
+      ))
       if (!is_html(url)) stop(
         "The file '", url, "' should have the '.html' or '.htm' extension."
       )
@@ -187,6 +193,7 @@ chrome_print = function(
       later::run_now(); if (!is.null(svr)) run_servr()
     }
 
+    if (is_rstudio_knit) message('\nOutput created: ', basename(output))
     invisible(output)
   })
 }
@@ -210,7 +217,8 @@ find_chrome = function() {
       res = head(res[file.exists(res)], 1)
       if (length(res) != 1) stop(
         'Cannot find Google Chrome automatically from the Windows Registry Hive. ',
-        "Please pass the full path of chrome.exe to the 'browser' argument."
+        "Please pass the full path of chrome.exe to the 'browser' argument",
+        "or to the environment variable 'PAGEDOWN_CHROME'."
       )
       res
     },
@@ -475,40 +483,4 @@ print_page = function(
   })
 
   ws$connect()
-}
-
-#' Convert Rmd to PDF with Chrome
-#'
-#' This is a helper function intended to be used with RStudio IDE. When the
-#' line \code{"knit: knit_pdf_chrome"} is present in the YAML header of an Rmd
-#' file, the behavior of the 'Knit' button of the RStudio IDE is modified: an
-#' extra PDF file is also generated using Chrome. This function is suitable for
-#' any R Markdown HTML output formats.
-#'
-#' If \code{knit_pdf_chrome()} cannot automatically find Chrome, use the
-#' environment variable \code{PAGEDOWN_CHROME} to set the path to Google Chrome
-#' or Chromium executable.
-#'
-#' @inheritParams rmarkdown::render
-#' @param encoding Not used. This argument is required by
-#'   RStudio IDE.
-#'
-#' @return Path of the output file (invisibly).
-#' @keywords internal
-#' @seealso \link{chrome_print}()
-#' @export
-knit_pdf_chrome = function(input, encoding) {
-  opts = options(pagedown.render.suppress.messages = TRUE)
-  on.exit(options(opts), add = TRUE)
-  dir = getwd()
-  output = chrome_print(input = input, verbose = 1)
-  message('\nOutput created: ', rmarkdown::relative_to(dir, output))
-  output
-}
-
-render_opt_msg = function(...) {
-  if (isTRUE(getOption('pagedown.render.suppress.messages')))
-    suppressMessages(rmarkdown::render(...))
-  else
-    rmarkdown::render(...)
 }